add IF_ARG macro conditionals
Good news: rgbds is not supposed to be able to export string macros
(blah EQUS ""), but it CAN! So I created some macro IF_ conditionals
which can replace a standard IF macro call. This allows for more
readable macro-branching based on IF-conditionals. The user still
has to end it with an ENDC, however.