rewrite math_Mult macro for speed
avoid using procedurall calls at all costs. For many #'s that can
be computed with computational shortcuts, use those shortcuts
in-line rather than call a procedure. This improves the speed of
most multiply calls (assuming 2nd arg is hard-coded or defined #)
by 25% or so. For your information: Call and ret take up a total
of 10 cycles, enough to x32 a number if no call is made.