add if_bit(s) and ldhard syntax
if_bits make it easier (reading-wise) to test for bits in a register.
The bit-testing is not limited to register a.
ldhard is a macro to ensure that the user passed in a defined # or
hard-coded # as an argument rather than a register. This is to
ensure that the program doesn't try to use a register after its value
has been overwritten, for example.
Also significantly improve load syntax error-printing.