add Iterator fxn and macros
(also change matrix _Y, _X to _H, _W)
Iterators must be declared, then Initialized before they can be used.
(declare reserves ram space, initialize sets them). Init also sets
the values for iterating the full matrix. Eventually macros should be
added to allow iteration over a submatrix within the full matrix.
IterNext returns with the next byte in A, and it's corresponding
address in HL. An IterStop is signalled by a cleared Carry-Flag (0).
IterStop indicates that you've already got the last element in the
previous iteration. So it's wise to "jp nc, xxxx" RIGHT after calling
mat_IterNext to avoid re-using the repeated value
IterStop will still return with the last element in A and it's
address in HL