add ifa_not
ifa_not covers cases where you'd like to reverse comparison
logic passed into your macro. This allows you to do it with minimal
fuss.
For example, if you have a macro, "array_count", where you count
the number of instances that match a comparison.
the user may call:
array_count	>=, 5
but you may need to reverse that comparison logic in order to do something
else if not >= 5. Within the "array_count" macro, \1 is ">=", \2 is 5.
To use ifa_not, you'd call:
ifa_not \1, \2, do_the_thing
where do_the_thing is the task to do IF the array element is not >= 5