rewrite IterInit and IterNext's counting
previously, we'd check if steps_remain == 0 before decrementing it.
If it was 0, we'd move onto stepping over to the next row.
Now, we decrement steps_remain, then check Z-flag. This has the benefit
of being faster and fewer bytes. But since it decrements and then checks
steps_remain, the loop will end one step earlier than before.
Therefore, in IterInit, we add 1 to both rowsteps_remain and steps_remain.
[bugfix] also add 1 more (again, I know) to the initial value of
steps_remain. This is because the first step moves the address HL from
outside the bounds of the matrix into the first byte of the matrix.
Then the value of steps_remain is equal to it's nominal value at the
first byte of each row.