fix bug: overflow when get/set probed cell status
after mat_IterNext is called, it returns with byte in A, and address
in HL. But this is the FULL ADDRESS. Aka Index + matrix_start_addr.
So we need to subtract start_addr to get Index. It's interesting
because the bug essentially made mat_GetIndex probed, hl
overflow and land probably in $A000+$A000 = $4000 range
(when you take into account the overflow / wrap-around).
Which means that I was accessing and setting random bytes within
ROM-space (which is $0000-$8000). Obviously, setting a byte would
have failed, and getting a byte would rarely equal 1. Which means
it was by pure chance that the stack didn't overfill.
But, whenever we pop an element from the stack, THAT probe setting
is correct, which is why the stack eventually finishes.