Initial debug module commit
the debug module is to be used while runnning a game. Simply
`call bug_snapshot` to write all registers AF, BC, DE, HL, SP, & PC
to the address $FF04 (rDIV). Register pairs are written MSB first
(so A before F, B before C, etc.). The routine bug_snapshot is written
such that it will report SP and PC as they were BEFORE bug_snapshot
was called. Additionally, it attempts to preserve and restore the
interrupt and timer state (as best as possible) so that any-timer
based interrupts / events will be minimally impacted.
You'll need to use custom hardware or modify an emulator to preserve
writes to $FF04 in a list in order to see the values.