check for dense mines only centered around a mine
we do this by skipping to next itation over the minefield
if we do not see a mine at the current location. An even
faster version might be to iterate over a list of mines,
but this way will work. This change caused a significant
drop in lag seen after pressing an initial button
-
Also add missing random.asm