improve color capability detection
previously, I'd detect if the hardware supported color, but
ran into an issue where I'd accidentally compiled the rom
without color support. But the game detected gbc hardware
and therefore attempted to nonetheless implement colors.
This led to the game being unplayable on the gbc (with a
black screen), but playable on the gameboy itself.
So now we check if the rom is also color-compatible