add dummy-check for which module is being compiled
too many times have I attempted to compile and run a submodule test
rather than the main test. This produces an unplayable rom, and bgb
complains about it's header and checksum. That should've been a dead
giveaway, but I've stumbled around this error long enough that a
reminder would be helpful