first pass at simulating multithreading
vblank processes all on-screen visual changes now.
It makes probing an area lightning fast. Small areas instananeously open
up, while larger areas get cleared in a few screen updates. It looks
amazing.
It displays all available revealed cells. In fact, it works perfectly
for that. But then it "freezes" once I probe a mine, or flag a cell.
Technically, it doesn't freeze since the crosshairs continue to respond
to my button-presses. But it is frozen. The mine-reveal and flagging
functions still use mat_SetIndex and/or lcd_Wait4VBlank, which no longer
triggers thanks to my very long v-blank handling routine. So it's
necessary to convert all functions that write to the screen to instead
write to a stack which gets pulled from during vblank. Then it'll work