first try at adding threads-unsafe argument
in order to prevent the re-enablement of interrupts. For example, during
a vblank, I make many calls to push and pop from stacks. If each one of
these enables interrupts (they are disabled during another interrupt),
then I run the risk of future interrupts causing havoc with my tight
VRAM timing requirements.
Note that this only works on pushing and popping 3 bytes at a time.
Which makes it a little unwieldy (and currently flagging / unflagging a
cell is still re-enabling interrupts because it only pushes / pops two
bytes at a time. Hence why I've set it to the bottom of the vblank
routine)