smoothly repeat movement if joypad held down
basically function like a keyboard where initial keypress causes a
movement of the player, and then if held down will continue to move in
that direction. But I quickly found that pressing Up+Left no longer
worked because it would move up, then wait the repeat-delay before
moving Up+Left. So I now detect when a new key is pressed during the
repeat-delay (called Charge) and move the extra direction before
resetting the charge so that the next repeat will happen a reliable time
after the last keypress.
--also--
wait for mines-stack to be depleted before adding red-mine. That way
it'll always get shown as red. (previously, the black mine may get
displayed after the red one, thus covering it up)