document how to throw lcd line-interrupt
basically, rIE needs to enable LCDC
and rSTAT needs to enable STATF_LYC
and rLYC needs a line-number written to it
and section "LCDC" [$0048] needs to be set up to handle the interrupt
~
Once all those are done, the LCD will trigger an interrupt when
rLYC==rLY, and $0048 will be called.
~
If you want to trigger an interrupt right before vblank, set rLYC to <140.
rLY range is 0-153. rLY == 140-153 during vblank, then it resets to 0