add CASE syntax
CASE does not behave like normal c-like CASE. CASE behaves like the
`ifa` macro except that after successful match and argument execution,
control will jump to label supplied in 3rd argument.
~
example:
    case    >=, 5, .endswitch, roll_ball 4
if A>=5, will execute `roll_ball 4`, and after-which it'll
`jp .endswitch`
If NOT A >= 5, then execution skips all of that and go to the next line.
(which can contain the next CASE statement)