add smoke particles after explosion
the smoke effect is achieved by loading a "circular" 8x8 tile with
transparency (by dithering) into ram. Each vblank, I alter a different
byte within the tile and AND it with the original tile. This keeps it in
the same shape and transparency as the original, while still causing
significant changes to the tile itself over time. The neat thing about
this is that the gameboy treats 1 bit from two sequential bytes as the
shading for one pixel in its tile. Meaning that by only changing one
byte at a time, the changes per-byte are more like shading changes,
rather than outright changes in structure. Gives the smoke effect a more
flowing structure.