move all non-vram updates from vblank to lcdc
previously lcdc was used exclusively for shaking the screen. Now it
shares responsibility for handling joypad, moving character, and queuing
flags. This is handy because it means that joypad movement and flagging
can happen 1 frame earlier (depending on timing) so that movement is a
little more responsive. Additionally, this frees up vblank to handle
only graphical updates, which makes its code more focused