add fast rotate macro
the rotate instructions RR, RRC, RL, RLC are 2 cycles & 2 bytes unless
it's specifically rotating the A register. Then its equivalent RRA,
RRCA, RLA, RLCA are all 1 byte, 1 cycle. I don't know if the RGBDS
compiler optimizes those instructions, so this macro is used to do
the optimization to calling the specific rotate if register A is used.