rewrite math_MultiplyAC for speed
this uses a (very confusing) trick from
http://sgate.emt.bme.hu/patai/publications/z80guide/part4.html
which is that we treat the rotation of bit 7 from A, and the
shift-left of BC as one single step: add HL, HL. If bit 7 from A
(now residing in H) is a 1, it'll set the carry-flag and we'll
"add HL, BC" which adds C to HL. But it doesn't mess with H
(synonymous with A) since B is always 0. It's weird
because Add HL, HL is both advancing a bit & shifting left,
except in the first ADD HL, HL. The first ADD only shifts bit 7
into carry, since L=0 at that time.