slow down firework explosion
move firework 1/16th distance to destination per frame. Then, because
this can lead to X & Y converging at different times on the destination,
explode the firework early if either X or Y matches destination. This
prevents the awkward "shift" of the firework as it homes in on the
destination coordinates when one coordinate is already satisfied.