fix subtle label-naming scope error
this fixes an error in my testing suite that I'd worked around 10 or so
commits back.
tests failed to compile because test_syntax.asm failed to find local
label .passed_0D
This error usually means that I mistyped the name of the label. For
example, passed_0D (forgetting the . in front of the label) will result
in a "passed_0D not found" error. However, my spelling was correct.
What I finally found was that the macro TestFailed accidentally defined
a non-local label
briefly_skip\@:
instead of
.briefly_skip\@
This unfortunate creation of a non-local label meant that TestFailed
changed the scope of all post-ceding labels; All local labels created
after TestFailed was called would be local to briefly_skip\@: instead
of the 0D_test scope.
Fixing the label meant that the call to jp .passed_0D now found the
local label which was still in-scope.
Interestingly, it looks like you can jump to a labels' locally-scoped
label by using '.'
such as test_0D.passed