add if_conditionals and asserts for macro args
In an effort to reduce duplication and simplify readable macro argument
verification, I've created a set of string-expansions to be used after
the IF conditional like so: IF arg1_IS_REGISTER.
Rather than: IF STRIN("A.B.C.D.E.H.L","\1")

Then since some macros just need to fail immediately if an argument
isn't a register (or something else) I've created the ASSERT_* macros,
which take an argument to verify and an error message (optional) to
display if the argument fails the assertion. For example:
`ASSERT_IS_NUMBER \2` will fail if \2 is not a number or a symbol that
equals a #. Use these assertions at the top of most macros to quickly
fail if an argument is out-of-spec. Wherever possible, I've used these
IF conditional string expansions and assertions in my code.